% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_best_model.R
\name{best_lm_model}
\alias{best_lm_model}
\title{Best Linear Model}
\usage{
best_lm_model(data, dep_var = "", keep_vars = list(),
  ignore_vars = list(), return = "all")
}
\arguments{
\item{data}{A \code{data.frame()} object which contains the data you wish to model.}

\item{dep_var}{A \code{string} containing the dependent variable you wish your model to measure.  Defaults to the first column of \code{data}.}

\item{keep_vars}{A \code{string}, or \code{list()} of \code{strings}, which represents the variables you want to keep in your final model regardless of their effects on the final model's optimization.}

\item{ignore_vars}{A \code{string}, or \code{list()} of \code{strings}, which represents the variables you want to exclude in your final model regardless of their effects on the final model's optimization.}

\item{return}{A \code{string} object that represents the type of model you would like to return.
\itemize{
 \item{"base" returns a linear model of all of the variables through the obect \code{model_base} and a summary table of that model.}
 \item{"r2" returns an Adjusted \eqn{R^{2}}{R^2} optimized linear model of the data through the obect \code{model_r2} and a summary table of that model.}
 \item{"p" returns a p-value optimized (95\% signifigance) linear model of the data through the obect \code{model_p} and a summary table of that model.}
 \item{"all" returns all of the above \code{model} objects and summary tables.  \code{return} defaults to "all".}
 }}
}
\description{
This function helps you select the best linear model given your dataset.  Allows you to define a denpendent variable, any variables you wish to have in your final model and which type of optomized liner model you would like returned.  Provides both Adjusted \eqn{R^{2}}{R^2} and p-value optimized models.
}
\examples(verbose=TRUE){
#best_lm_model(swiss, dep_var="Fertility", keep_var="Catholic", return="p")
print("Hello world!")
}